+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 Yii Framework 2 bootstrap4 extension Change Log
 ==============================================
 
-2.0.7 under development
+2.0.9 under development
 -----------------------
 
-- no changes in this release.
+- Enh #191: Added `centerVertical`, `scrollable`, `dialogOptions` options to Modal and `role="document"` to `modal-dialog` element (adhayward)
+
+
+2.0.8 October 08, 2019
+----------------------
+
+- Bug #177: Fixed non-functional headerOptions in Tabs (Thoulah)
 
 
 2.0.7 August 13, 2019